refactor: move Red Hat fetch logic to External Access layer
Move _fetch_redhat_doc, _get_redhat_client, _estimate_tokens, and
client globals from docs.py (Domain) to redhat_docs.py (External
Access). Route docs.redhat.com URLs from server.py (Orchestration)
directly to redhat_docs.fetch_redhat_doc — Orchestration is allowed
to call External Access, resolving the Domain → External Access
layer violation.

Also adds response size bound (2MB) in _call_tool and sanitizes
MCP error messages via sanitize_error().
